+++ Mods ChoGGi/1 Percenter/metadata.lua
@@ -3,15 +3,15 @@ return PlaceObj("ModDef", {
 	"id", "ChoGGi_1Percenter",
 	"steam_id", "1411216506",
 	"pops_any_uuid", "63bbd317-5310-4f6c-8ff4-24e715812c1e",
-	"version", 3,
+	"lua_revision", 1001514, -- Tito
+	"version", 4,
 	"version_major", 0,
-	"version_minor", 3,
+	"version_minor", 4,
 	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua"
 	},
 	"image", "Preview.png",
-	"lua_revision", 1001514, -- Tito
 	"description", [[Receive a bonus of one percent of your funds each Sol.
 
 Requested by: EasyMoney]],