+++ Mods ChoGGi/Custom Resource Icon Toggle/metadata.lua
@@ -1,0 +1,24 @@
+return PlaceObj("ModDef", {
+	"title", "Custom Resource Icon Toggle",
+	"id", "ChoGGi_CustomResourceIconToggle",
+	"steam_id", "2473788321",
+	"pops_any_uuid", "950f7c05-1199-4c72-82bd-adad9edb3e02",
+	"lua_revision", 1001514, -- Tito
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagCosmetics", true,
+	"TagInterface", true,
+	"description", [[
+Mod options to selectivly toggle resource/effect icons (pressing "I" will reset it, re-apply mod options).
+
+
+Requested by ethan.
+]],
+})