+++ Mods ChoGGi/Fix Flying Drones MidAir Malfunction/metadata.lua
@@ -1,0 +1,20 @@
+return PlaceObj("ModDef", {
+	"title", "Fix Flying Drones MidAir Malfunction",
+	"id", "ChoGGi_FixFlyingDronesMidAirMalfunction",
+	"steam_id", "2473394779",
+	"pops_any_uuid", "88b79b2c-a54b-4f9b-8d58-9e69e03def3b",
+	"lua_revision", 1001514, -- Tito
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagOther", true,
+	"description", [[
+Flying drones that malfunction mid-air are stuck "flying".
+]],
+})